+++ docs/history.md
@@ -45,4 +45,5 @@ development started from 1.1.2024
 
 ## 2024 week7
 * scrape web
+* lazy loading (only addresses that are visible are loaded)